adding orchestrator consumers refactoring

+++ services/ochestrator/src/main/java/com/graduationproject/ochestrator/saga/SagaParticipators/CreateResident.java
@@ -17,8 +17,8 @@ import org.springframework.stereotype.Service;
 import javax.transaction.Transactional;
 import java.util.UUID;
 
-import static com.graduationproject.ochestrator.topic.ResidentTopics.CreateResidentSagaBegin;
-import static com.graduationproject.ochestrator.topic.ResidentTopics.CreateResidentSagaFailed;
+import static com.graduationproject.ochestrator.topic.resident.ResidentTopics.CreateResidentSagaBegin;
+import static com.graduationproject.ochestrator.topic.resident.ResidentTopics.CreateResidentSagaFailed;
 
 @Service
 public class CreateResident implements SagaParticipator<ResidentDto> {
@@ -65,8 +65,12 @@ public class CreateResident implements SagaParticipator<ResidentDto> {
     @Transactional
     public void transact(String sagaId) {
         //this will be run after a successful saga
+        Resident resident = residentRepository.findResidentBySagaId(sagaId);
         residentRepository.deleteBySagaId(sagaId);
-        departmentRepository.deleteBySagaId(sagaId);
+        if (residentRepository.countByDepartment(resident.getDepartment()) == 0) {
+            departmentRepository.deleteBySagaId(sagaId);
+        }
+
     }
 
     @Transactional
